coproc/js: add awaitWriteFinish to chokidar js
By default chokidar sets the value for the watch parameter
'awaitWriteFinish' to false. This means the library will trigger
the onWrite callback before the entire file has been written to
disk. When testing with redpanda it was observed the the wasm engine
 attempted to open and read a file that was partially written
to disk. The result was an exception from the use of this
erroneous file.

By settings 'awaitWriteFinish' to true, we guarantee that
our watcher callbacks will only be called when files are
fully written to disk, removing any race conditions
that could occur.