coproc: Free functions for validating and parsing wasm_events
- There are two types of wasm events records with keys and values (deploy
events) and records with just keys (remove events)

- Remove events just contain the name of the coprocessor to remove. This is in
the 'key' field.

- Deploy events also contain this name in the 'key' field, however it contains
the actual javascript source code of the coprocessor in the 'value' field. Also
there is some metadata stored in the record_headers vector. Description,
action (explicity defining the action type) and 'sha256', a checksum of the
source code.

- The functions try to keep the amount of data copied to a minumum, only
performing copies when neccessary.