coproc/tests: Generators for wasm_events
- There are a few different methods exposed in the header file
'wasm_event_generator.h' to support the different types of test cases.

- make_wasm_event_record_batch_reader, for streaming random batches.

- create_wasm_record(...) and the wasm_event type are for producing one record
with the desired fields. All fields in the wasm_event struct are optional, if a
value is std::nullopt it will not be serialized in the resultant record. The
idea is to produce potentially invalid records and ensure that the validators
fail with the correct error.

- make_wasm_batch(offset, vector<record>) this method is for making a
record_batch of distinctly produced records from the create_wasm_record above.
Useful when its desired to string together multiple definitive events to
generate a partiticular final tests state.