rpk: send -1 as create topics default replica count
The patch
https://github.com/vectorizedio/redpanda/commit/144264580086025982f9d146a7634f64675b3f09
changes the default replica count to be -1 which requests the server to
a use a default value. however, when the configuration object is
constructed before handing it off to sarama the replica count field is
not set if the value is less than zero (-1 < 0) and the result is that
the server receives requests with a default replica count of 0.

Signed-off-by: Noah Watkins <noah@vectorized.io>