k/client: remove dependency on supported api range
The existing kafka client sent requests to a broker with the same max
supported api version as the redpanda server supports. This is
problematic for couple reasons. First, the features that our client
supports is going to be evolving at a different rate than the server and
should perform a form of version negotation with the server via the api
versions broker interface. Second, we may soon find examples of where it
is convenient to reuse our client to talk to non-redpanda brokers (e.g.
data migration scenarios).

To address this I've removed the dependence on the max supported API and
retained the existing set of values / behavior in the kafka client which
can be removed once version negotiation is implemented.

Signed-off-by: Noah Watkins <noah@vectorized.io>