s/consensus: add initial revision to config coming from snapshot & log
Adding revision to revision less configuration that were stored in raft
snapshot by previous version of redpanda.

Similarly to #504 the configuration stored in snapshot metadata may be
in old format. We have to assign initial revision to the configuration
stored in snapshot metadata before adding it to configuration manager as
it may already exists at the same offset. If configuration stored in
snapshot has the same offset as the one in `raft::configuration_manager`
it means the configuration have to be identical. The same rule apply to
the configuration stored in raft log. Whenever it configuration is
revision less we have it must be assigned with initial revision.

Signed-off-by: Michal Maslanka <michal@vectorized.io>