coproc: Fix for crash on script_manager double delete condition
- The root cause of this error is that the coproc::service erranously believes
that v/rpc servers are not concurrent with respect to 1 TCP connection. This is
not true, as the rpc server issues the next request after the body of the
preceeding request has been parsed.

- Because of this its possible for two calls of remove_source to interleave so
that shutdown() is called twice on a script_context, causing the actual crash.

- The solution is to remove the script context from the _scripts collection as
the first item of business in the remove_source method. Since this does not
occur in a future this action cannot be reordered and subsequent calls to
remove_source correctly report that the item does not exist in the cache

- Fixes #538