r/replicate_stm: skip sending requests to not responding follower
When follower is down, depending on the configuration raft group may be
able to still operate as it has majority. In the case when follower
requests timeout (f.e. network partition, follower process is
suspendend) we have to propagate backpressure to replicate_entries_stm
but still allow replication to healthy nodes. In order to do this
without compromising safety and performance we can skip sending append
entries request to not responding followers.

When follower is not responding it will require to be brought back up to
date after failure recovery. Skipping sending append entries requests
prevent requests queue build up and relief memory pressure.

Signed-off-by: Michal Maslanka <michal@vectorized.io>