tests: add topic specification in ducktape
The redpanda ducktape test harness has a convenience parameter to the
constructor for creating initial topics. This interface can be
inconvenient because it requires overloading the constructor in the test
subclass.

This patch adds a TopicSpec object which is a collection of topic
specifications (like name and configuration) and allows initial toipcs
as set of TopicSpecs to be specified as a class variable without
overloading the constructor.

Signed-off-by: Noah Watkins <noah@vectorized.io>