gcb: enable CCACHE for build and build-pkg pipelines
Uses https://github.com/GoogleCloudPlatform/cloud-builders-community/tree/master/cache
to save/restore ccache to/from a GCS bucket. The content of the cache is
placed in /dev/shm and the CCACHE_DIR variable is set from the pipeline
so that it is properly picked by vtools. In the best-case scenario, this
saves ~15 mins on a release build.

The cache image was built once (manually), and uploaded to the
vectorized private GCR registry. In addition, the GCS bucket was also
manually created.

Signed-off-by: Ivo Jimenez <ivo@vectorized.io>