Enabled ccache on the GitHub Actions CI workflow
Speeds up a full build from 33 minutes (with unity builds) down to 7 minutes
when the cache is warm - that would happen after the first build of a branch
when the cache is initially created. Unfortunately caches cannot be reused
between branches even though the cache action documentation says otherwise.
Also using the build matrix container in the cache keys.