archival: Add archival scheduler service
The service is responsible for S3 uploads. It wires up with log_manager
and partition_manager. Periodically, it checks what partitions are
available and adds them to the queue. It fetches a bunch of sealed
segents from the queue periodically and uploads them. Only leader for
a particular NTP uploads segments that belongs to that NTP.

Also, service is responsible for uploading topic manifests.
Service never deletes anything from S3. Partition reconciliation
runs periodically and upload runs in the background.