test: add ducktape redpanda service without client
When the redpanda service starts it verifies that the cluster is healthy
before proceeding. This check is done with a client that doesn't yet
support sasl configuration. This patch adds a specialization of the
service that skips this check. Test users can perform the check
themselves or rely on the check being implicit by failing to communicate
with the cluster.

This is largely temporary until we have better SASL support.

Signed-off-by: Noah Watkins <noah@vectorized.io>