operator/connectivity: Add external dns resources
The statefulset now exposes each pod using the host port defined in
kafka external port. Based on the pod placement the public IP address
of an node is taken to register in DNS provider.

Operator exposes redpanda nodes using the exteranl dns project.
For the schema the operator adjust it to vectorized.io convencion.