(split) operator/tests: move artifacts to a different folder
e2e folder is folder that kuttl uses to looks for tests, by having artifacts in the same folder, repeated local runs causes that the folder is interpreted as another test that is empty and skipped. This prevents kuttl from needing to parse that folder as we move it one level up.