storage: fix batch cache instance construction in log messages
The batch cache reclaim_options did not have a output operator defined
for it. But the batch cache did.

When this batch cache constructor was not explicit:

   explicit batch_cache(const reclaim_options& opts)

Then this output stream operator for log config:

   std::ostream& operator<<(std::ostream& o, const log_config& c) {
      ...
      << ", relcaim_opts:" << c.reclaim_opts << "}";

Was constructing a batch cache object from the reclaim options and
printing it out instead of the options. Basically harmless, but still it
was incorrect and would register then immediately unregister a reclaimer
with seastar.

Fix is to define an output operator for reclaim options and mark the
constructor as explicit.

Signed-off-by: Noah Watkins <noah@vectorized.io>