rpc/transport: keep socket alive after connection timeout was triggered
Fixed removing socket after timeout was triggered but connection was
eventually created. In this case we can not remove the socket structure
until the future is finished. When timeout occurs, we shutdown all
connection attempts and already created connections so even if
connection was established it will be safely closed.

Signed-off-by: Michal Maslanka <michal@vectorized.io>