coproc/js: implement new deregister and skip strategy in RPC
those changes modify the ProcessBatchReplyItem type, now
the record attribute is a Optional<Array>.

what does it change?
now error policy is going to handle via record value,
for example, if one coprocessor script fails and that
coprocessor has errorPolicy=Deregister, the record
value in ProcessBatchReplyItem is going to be `undefined`,
when redpanda validate that record value, send a
disableCoprocessor request to wasm-engine.

for skipOnFailure policy the record value in ProcessBatchReplyItem
is going to be a empty array.