ts-generator: fix writtenBytes variable
this variable shouldn't add to the template if the `write_array`
isn't the first type on origin type, for example `Optional<Array>`,
`writtenBytes` should add if the original type is `Array<Optional>`

what is the problem? this variable was adding one increment
operation on that function, for that reason, the result value
had more bytes than expected

Why is this error happening now? Before we didn't have Array<T>
as a value type, for example: Optional<Array<T>>