coproc: Use a plain kafka::transport within the copro::event_listener
- This kafka::client will connect to every node in the cluster. Since every node
has 1 instance of an event_listener this will cause every node to connect to
every other node, which isn't desired.

- Long term this behavior with the client is to be dropped in favor of having
the client only connect to brokers on a need-to-know basis.

- This also resolves an issue observed in the partition_move_test during
cluster restarts. The issue stems from the fact that redpanda returns 5 nodes
during a metadata request after reboot, which is the last observed state of the
world, and eventually will be consistent with the current state. The client
attempts to connect to all 5 and throws, without proper error handling from the
event_listener redpanda crashes.

- Swapping the kafka::client for a kafka::transport isn't enough to prevent this
same issue from happening here. Proper error handling must be introduced and it
has been. So in the event a broker cannot be reached, we wait 1s and begin the
do_connect produre again.

- Given the two points above it was decided to swap out the kafka::client for
the simpler transport mainly because the event_listener should consider progress
made only after the single broker containing the coproc_internal_topic can
be reached. It shouldn't have to wait until all previously connected brokers
also can be reached.

- Fixes #705