cluster: fixes a data race on id_allocator_frontend initialization
When Redpanda creates a topic and immediately uses it it leads to
a problem because an initialization of the internal structures such
as ss::sharded<shard_table> happens after acknowledging a request
and it's a race.

The change adds a retry logic not to fail the first request to
id_allocator_frontend when we create a topic on demand.