archival: Make endpoint configurable
GCS and other S3 implementations use different endpoints and different
url format (e.g. no region in the url). This commit adds configuration
option that can be used to override the endpoint url. In this case
the automatic AWS endpoint selection is disabled and 'region' parameter
is ignored. The user is supposed to provide the right endpoint for the
region that he/she uses.