model: add producer compression type
Adds an enumeration value for `producer` compression type which is a
topic/broker level configuration option that states batches should
retain the compression type used by the producer. Since this was
effectively the default behavior before this change but we rejected
configuration value of `producer` this change only needs to accept the
value during topic creation.

Note that producer type is not a batch compression type it is a policy
on a topic to retain the producer's original compression type.

Signed-off-by: Noah Watkins <noah@vectorized.io>