cluster/stm: Adds persisted_stm::sync to wait until a state machine catches up
Adds persisted_stm::sync method to wait until a state machine catches up with
a given offset. persisted_stm::sync is very similar to state_machine::wait but
it instead of throwing an exception on timeout it returns a boolean.

The motivation is to avoid exceptions to simplify the coroutine code.