kafka: take request_context by value in server handlers
As we move towards more pervasive use of coroutines we need to be very
careful about how we treat coroutine parameters, in particular reference
parameters.

1. reference parameters in coroutines _look_ far more safe because the
   coroutine looks like normal blocking code. since we are mixing styles
   this is usually not true.

2. taking reference parameters by value in coroutines treats them
   similarily to what they are in reality: lambda captures. this also
   eliminates the need for hacks to keep copies alive on the stack:

      task foo(type&& ref) {
        auto copy_that_stays_alive = std::move(ref);
        ...
        co_return;
      }

  the parameters themselves are kept alive for the coroutine lifetime.

This patch makes this change for request_context parameter to kafka
server handlers. This type is small and already often moved into a
`do_with` block to keep it alive. This change prepares our kafka
handlers for easier time with transition to coroutines.

Signed-off-by: Noah Watkins <noah@vectorized.io>