operator: when subdomain is provided, don't pull node
The node object is not used when subdomain is provided so we should not be
querying it. This surfaced during testing in kind when nodes does not have name
provided.

ERROR	controller-runtime.manager.controller.cluster	Reconciler error
{"reconciler group": "redpanda.vectorized.io", "reconciler kind": "Cluster",
"name": "cluster-tls", "namespace": "helm-test", "error": "failed to construct
external node list: failed to retrieve node : Node \"\" not found"}