kafka: return error code from kafka list groups api
Prior to this change we returned a bool indicating if the coordinator
was loading (this was the only error possible) and converted the bool
into the error code when constructing the response.

This change returns an error code which simplifies the handler which
builds the response and makes it easier to return other error codes from
the group manager.

Signed-off-by: Noah Watkins <noah@vectorized.io>