rpk/topic/create: Rename --config to --topic-config to fix conflict
`rpk topic` has a global --config flag which, like in the rest of
commands, points the command to the desired config file path.
However, `rpk topic create` declared another flag named --config,
but for passing custom k:v pairs to be sent as the new topic's
configuration. Rename the latter to --topic-config to fix the
conflict between the two.