kafka: add helpers for acl state conversion
Internal ACL management in Redpanda is nearly identical to the data
model used in Kafka, but we have chosen to completely separate the
internal implementation from the details of the Kafka API in
anticipation of divergent requirements as we will be integrating ACLs
into other parts of Redpanda like the proxy and admin http api in
addition to other services.

This patch adds details for converting between the two abstractions.

Signed-off-by: Noah Watkins <noah@vectorized.io>