security: update store binding removal
Previously the binding removal was brute-force and did not seem to
handle all of the scenarios that were accounted for in the kafka removal
process. This patch updates the algorithm to largely be 1-1 with the
strategy / semantics found in kafka's removal process. Much of the
complexity arises from what is needed to support the kafka api, namely
that the removal response contains the bindings that were removed in the
same order as the input filters that were used to match the acl entries
that resulted in the binding removal.

Signed-off-by: Noah Watkins <noah@vectorized.io>