redpanda: Fix kafka server and rpc server intiialization
The problem: we introduced multi-listeners and per-listener credentials
to redpanda. To do that 'server_endpoint' class was introduced. The
instances of this class contains shared_ptr<server_credentials> field.
This field was initialized in main and passed to the shards. The
server_endpont was copied and this triggered assertion in shared_ptr
since it was created on different shard.

This PR adds c-tor to rpc::server that accepts ss::sharded<configuration>
and uses local() to obtain its shard-local copy of the config. Also, it
initializes rpc::configuration for both kafka and rpc servers as a
sharded<rpc::configuration> so the actual initialization and allocation
of the server_credentials is shifted to the shard.