archival: Enable minio on CI
Update docker-compose.yml to run minio image in a useful way. The
environment contains a bunch of variables which are used in tests
e.g. credentials and region name. Also it has MINIO_DOMAIN var which
is needed for redpanda (it enables virtual host style path in requests
for REST API).