kafka/client: Introduce partitioners
A partitioner takes a record and assigns it a partition.
A partitioner may fail to assign a partition, so it returns an optional.

Several strategies are included:

* Identity - return the partition id already in the record
* Murmur2 - assign a partition id dervied from the murmer2 hash of the key
* roundrobin - Use round-robin to assign a partition id
* default_fallback - Use all three, in order. Cannot fail.

The default partitioner in java client uses a different strategy, with
a sticky partitioner where round-robin is. The sticky partitioner
moves to the next partition when the batch is full. This strategy should
be implemented later.

Signed-off-by: Ben Pope <ben@vectorized.io>