kafka: add rw locks to group manager
Processing of the transactional requests must depend on the latest state
and when it becomes outdated (caused by lost leadership and out of band
update) may lead to incorrect results. The commit introduces a rw lock
to prevent the group manager from processing transactions when the
replay is happening.

We already have the p->loading flag to reject requests during the reload
proces but this mechanism doesn't drain the ongoing requests so there is
still a chance for the reload process to interleave with the request
processing.