kafka: add batch wrapper for codegen producer types
Code generation for produce request uses an iobuf for the kafka record
batch data. However, once it is read off the wire it needs to be
converted into a native batch type, which might in turn also fail (e.g.
bad checksum). Therefore we need to represent in the generated code
additional state beyond the native iobuf type used to hold raw data
on/off the wire.

This patch adds a wrapper type that handles the conversion when decoding
off the wire, and holds a normal batch which can also be set when
encoding a request. When encoding an overload in the response writer
handles this specific type.

The code generator simply replaces iobuf with this wrapper type in the
generated structures.

Signed-off-by: Noah Watkins <noah@vectorized.io>