s/readers_cache: implemented readers cache
Implemented cache for `storage::log_reader`. The cache holds reader
instances and allows user to query for reader using reader
configuration. If any of the readers kept in a cache matches given query
its configuration is reset and it is returned to the caller. Caller is
responsible for adding readers to the cache using `put()` method. Since
readers keep read lock to underlying segments `readers_cache` exposes
interface to force readers eviction in face of truncation and segments
removal. Readers are evicted from the cache according to LRU policy and
automatically when they can not longer be reused (f.e. EOF).

Signed-off-by: Michal Maslanka <michal@vectorized.io>