kafka: track and report client host and id per member
introduced client_host and client_id as strong types for per consumer
group member data. the host is represented as a string so that it may be
easier interoperate with outside tools that inspect consumer group data
directly from topic data.

since member metadata is also serialized to disk the on disk format must
be updated.

Signed-off-by: Noah Watkins <noah@vectorized.io>