kafka: add group::prepare_tx handler
prepare_tx is a handler for new server <-> server RPC of the Redpanda
tx protocol. It writes (prepares) the cached offsets of the inflight
transaction to disk. Before writing to disk it checks that current
term matches the term reported by begin_tx to be sure that a reboot
hasn't happened and hasn't invalidated cached offsets