s/parser: reusing already consumed record batch header
When batch parser is reused we have to deal with the fact that consumer
can decide based on the record batch header if parsing underlying data
source should be stopped. When stop is requested after reading batch
header from `input_stream`, parser have to keep the header in its
internal state. Header can not be reset to initial state as reading
header once again would lead to corruption since there are record data
still in the `input_stream`. Fixed this issue by resetting parser state
only when whole batch was consumed and skipping reading header again
when it is cached in parser state.

Signed-off-by: Michal Maslanka <michal@vectorized.io>