r/tests: reduced number of nodes in replace group test
Reduced number or concurrently running raft instances in
`replace_whole_group` test. In debug mode the test was failing in CI as
all of the raft instances are created on the same core, reduced number
or raft instances from 6 to 4 to minimize overhead and prevent test from
generating false failures.

Signed-off-by: Michal Maslanka <michal@vectorized.io>