http: Add monitoring probe
The probe counts input/output bytes, number of requests (get/put/others)
and transport errors. It doesn't use probe from RPC package because of
the specifics (e.g. with http we have long running requests that upload
a lot of data at once and there is a lot of transient/parallel
connections to the same host).