coproc: Copro heartbeat mechanism returns num of registered scripts
- This way a periodic evaluation of the internal state of the coprocessor engine
can be performed. If the expected state doesn't match what redpanda thinks it
should be, the failure recovery mechanism can begin.

- More advantageous to do this vs the old impl which would just consider
failures if a heartbeat wasn't heard within a timeout interval. This meant that
quick restarts of the wasm engine would incorrectly cause a state disruption
which redpanda would fail to notice.

- Fixed #1140