utils: add queue depth control algorithm
Adds an emwa control algorithm to manipulating the queue depth of a seastar
semaphore with the goal of maintaining latency below a given threshold.

This is loosely based off of the control algorithm in PARDA FAST '09 but
without any of the proportional allocations or distributed sharing of
state. It only uses the most basic form of queue depth control.

Signed-off-by: Noah Watkins <noah@vectorized.io>