archival: Fix CI issue
Originally, the archival test suite created a topic with two partitions
and used kafka cli client to generate data. For some strange reason the
client sometimes generates skewed workload. As result, segments from
one of the partitions are not uploaded and this leads to failure. This
commit takes number of partitions down to one. This still allows test
to verify the system without this problem getting in the way.

Another problem that this PR solves is MinIO consictency gap.
ListObjects REST API call in S3 is eventually consistent. MinIO is no
different. When the PutObject request report success ListObjects stil
wouldn't return it for a while. This sometines leads to test failures
due to timeout. This commit introduces _list_objects function that
uses GetObject REST API request to generate object list (by fetching
the manifest). The GetObject request provides read-your-own-writes
guarantee so it doesn't introduce this problem. Also, the regullar
ListObjects request is used as a fall back mechanism.