r/consensus: fixed reporting under replicated replicas
Fixed reporting under replicated replicas when node is down. The problem
was caused by incorrectly assuming that when node is under replicated it
is recovering. This assumption is incorrect for nodes that are down,
since recovery is triggered after receiving response from the node that
needs to be caught up.

Fixes: #1219

Signed-off-by: Michal Maslanka <michal@vectorized.io>