kafka: generate extra context field members
This feature in the code generator allows some extra fields that are not
part of the wire protocol to be added to the generated structured. This
is useful for adapting the fetch request because some state was tracked
at the partition level. An alternative approach would require tracking
this some where else like in a parallel data structure or index adding
complexity and overhead.

Signed-off-by: Noah Watkins <noah@vectorized.io>