kafka: take batch reader from lval ref in writer
Taking as an lvalue ref is a trade-off. It doesn't affect
performance--we still move from the value--but in return we do not have
to either (1) annotate all of the generated types to use std::move or
not, or alternative (2) use std::move for everythign even though most
types are trivially copyable.

Signed-off-by: Noah Watkins <noah@vectorized.io>