rpk/admin-api: Make requests to all nodes
As of v21.4.15, the Redpanda admin API doesn't do request forwarding, which
means that some requests (such as the ones made to /users) will fail unless
the reached node is the leader. Therefore, a request needs to be made to
each node, and of those requests at least one should succeed.

This issue is being tracked in https://github.com/vectorizedio/redpanda/issues/1265