s/disk_log_impl: determine compaction state in first compaction cycle
Using `storage::compaction_result` to check if the segment was actually
compacted or it didn't require compaction. When segment was already
compacted we skip to the next one. This way we establish segment
compaction state in single compaction cycle.
Previously, after each restart, we recognized single segment state
per compaction cycle. This ways with default settings and 10 segments in
a log it took 50 minutes to start compacting segments.

Signed-off-by: Michal Maslanka <michal@vectorized.io>