coproc: Do not deregister script if some subscriptions dne
- It is normal to expect that a script subscribed to a topic, may not observe
that any ntps for that topic exist on another node. For these cases, the script
should not be deregistered.

- Only immediately deregister a script if none of its subscriptons exist on a node.

- The example that made the bug obvious was a 3 node cluster, 3 topics, 1
partition for each topic with all topic replication factor set to 1. It was the
case that every script had at least one subscription that didn't exist on its
respective node, therefore all scripts were deregistered, and no processing
occurred.