archival: Fix reupload bug
Reupload wasn't triggered in case of timeout. The bug was caused by
the design problem in the connection pool. During reconnection the
abort_requested_exception was thrown on timeout. This caused the
ntp_archiver to skip the upload. The exception type was changed and
the reconnection loop works as intended.
Also, commit extends logging to improve observability.