kafka: handle requests in-order until auth completes
until authentication is complete, process requests in order since all
subsequent requests are dependent on authentication having completed.

the other important reason for disabling pipeling before authentication
completes is because when a sasl handshake with version=0 is processed, the
next data on the wire is _not_ another request: it is a size-prefixed
authentication payload without a request envelope, and requires special
handling.

a well behaved client should implicitly provide a data stream that invokes this
behavior in the server: that is, it won't send auth data (or any other
requests) until handshake or the full auth-process completes, etc... but
representing these nuances of the protocol _explicitly_ in the server makes its
behavior easier to understand and avoids misbehaving clients creating
server-side errors that will appear as a corrupted stream at best and at worst
some odd behavior.

Signed-off-by: Noah Watkins <noah@vectorized.io>