kafka: support handshake v0 sasl authentication
When SASL handshake v0 is encountered during authentication phase the
connection is expected to then not have a sasl auth request but rather
the raw auth token. Therefore, we add handling to the protocol state
machine that intercepts the connection data in this case and proxies the
raw authentication tokens to request processing by constructing
synthetic requests and then unwrapping them to pass back to the client.

Signed-off-by: Noah Watkins <noah@vectorized.io>