build: allow embedded git version to be disabled
Passing ENABLE_GIT_VERSION=OFF to cmake when configuring redpanda causes
the embedded git version information to be static, avoiding the need to
re-link every executable after the git HEAD changes. This is an
optimization for developer environments and should remain turned on when
building in CI or building releases. It is on by default, but only need
to be set when a build configuration is bootstrapped.

Signed-off-by: Noah Watkins <noah@vectorized.io>