rptests: Add a lock to restrict access to the config file
In the rpk produce/ consume tests, an async consumer service is
used (RpkConsumer, in services/rpk_consumer.py), they both use
RedpandaService#brokers, which opens and loads the nodes'
redpanda.yaml config file.

In 99.99999% of the cases, this creates a deadlock and the tests
hang, which is solved by locking access to the config file.