kafka: clear unprepared transactions on re-election
Redpanda caches ongoing transactions in memory and flushes them to disk
during the prepare phase of the commit protocol. To prevent data loss
caused by restart of a node transaction coordinator commits a transaction
only if the leader doesn't change during the duration of the transaction.

So it's safe to wipe out all ongoing unprepared transaction on re-election
to safe memory since the removed transaction anyway can't be comitted.